+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * [Enhancement] Allow attaching to the winches surface to let reinforcing them with struts.
 * [Enhancement] Don't show resources that cannot be transferred in the RTS GUI.
 * [Enhancement] Allow disabling the controls hints in the RTS resource transfer dialog. Use setting `showTransferDialogHints`.
+* [Enhancement] Allow disabling the controls hints in the winches remote control dialog. Use setting `showRemoteControlDialogHints`.
 * [Enhancement #248] Add ability to scale Transfer GUI.
 * [Enhancement #321] Scale the Winch GUI dialog.
 * [Fix #302] GUI does not respect hide/show function.